% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFB.R
\name{addChroniquesFile}
\alias{addChroniquesFile}
\title{Chronique file to optimisation output}
\usage{
addChroniquesFile(outputName)
}
\arguments{
\item{outputName}{\code{character}, name of output directory}
}
\description{
Chronique file to optimisation output
}
