% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColumnTypicalDay.R
\name{addColumnTypicalDay}
\alias{addColumnTypicalDay}
\title{Add typical day columns}
\usage{
addColumnTypicalDay(data, opts = antaresRead::simOptions())
}
\arguments{
\item{data}{\code{antaresdata} data load by \link{readAntares}}

\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. Defaut to \code{antaresRead::simOptions()}}
}
\description{
Add typical day columns
}
\examples{

\dontrun{

antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/antaresStudy", 1)
data <- readAntares(mcYears = 1:10)
data <- addColumnTypicalDay(data)


data <- readAntares(areas = "all", links = "all", clusters = "all" ,mcYears = 1:10)
data <- addColumnTypicalDay(data)

}

}
