% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNetPosition.R
\name{addNetPosition}
\alias{addNetPosition}
\title{Add net position}
\usage{
addNetPosition(data, opts = antaresRead::simOptions(), inAreas = c("be",
  "de", "fr", "nl"), ADQ = FALSE, newName = "_CWE")
}
\arguments{
\item{data}{\code{antaresDataList} read with readAntares and with areas and links tables}

\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. Defaut to \code{antaresRead::simOptions()}}

\item{ADQ}{\code{boolan} apply straitemùent on ADQ columns}

\item{newName}{\code{character} end of new columns name. Default "_CWE".}

\item{opts}{\code{character} areas who will included in IPN compute}
}
\description{
Add net position
}
\examples{
\dontrun{
opts <- antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/antaresStudy", 2)
data <- readAntares(area = "all", links = "all", mcYears = 1)
data <- addNetPosition(data, opts, ADQ = FALSE)

}

}
