% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFB.R
\name{addTSFile}
\alias{addTSFile}
\title{Time Series file to optimisation output}
\usage{
addTSFile(outputName)
}
\arguments{
\item{outputName}{\code{character}, name of output directory}
}
\description{
Time Series file to optimisation output
}
