% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColumnTypicalDay.R
\name{addTypicalDayId}
\alias{addTypicalDayId}
\title{Add typical day columns}
\usage{
addTypicalDayId(data, fb_opts = antaresRead::simOptions())
}
\arguments{
\item{data}{\code{antaresdata} data load by \link{readAntares}}

\item{fb_opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath} or fb model localisation obtain with \link{setFlowbasedPath}. Defaut to \code{antaresRead::simOptions()}}
}
\description{
Add typical day columns
}
\examples{

\dontrun{

antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/antaresStudy", 1)
data <- readAntares(mcYears = 1:10)
data <- addTypicalDayId(data)


data <- readAntares(areas = "all", links = "all", clusters = "all" ,mcYears = 1:10)
data <- addTypicalDayId(data)

}

}
