% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adqPatchH5.R
\name{adqH5}
\alias{adqH5}
\title{Run the adequacy patch and save the result in an h5 format file}
\usage{
adqH5(opts, fb_opts, strategic_reserve_be = NULL,
  strategic_reserve_de = NULL)
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. By default, the value is given by \code{antaresRead::simOptions()},
it must refer to an h5 file.}

\item{fb_opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath} or flow-based model path obtained 
with the function \link{setFlowbasedPath}. By default, the value is returned by \code{antaresRead::simOptions()}.}

\item{strategic_reserve_be}{\code{character}, name of the virtual area representing the strategic reserve of Belgium}

\item{strategic_reserve_de}{\code{character}, name of the virtual area representing the strategic reserve of Germany}
}
\description{
This function runs the adequacy patch (\link{adqPatch}) on an Antares output and saves the result in an h5 file.
}
\examples{

\dontrun{


##Opts defind on txt files
opts <- antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/antaresStudy",2)

##Write h5 stud
antaresRead::writeAntaresH5()

##Do a setSimulationPath on h5 file
opts <- antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/Antares/antaresFlowbased",1)

##Use for user folder, you cal also use setFlowbasedPath
fb_opts = antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/antaresStudy", 2)

##Compute adq on H5 file
adqH5(opts, fb_opts)

##Set alias for an easy read
setAlias("adqPatch", "Alias for adqPatch", c("LOLD_ADQPatch",
                                             "UNSP. ENRG_ADQPatch", "DTG MRG_ADQPatch",
                                             "UNSP. ENRG_ADQPatch", "BALANCE_ADQPatch",
                                             "FLOW LIN._ADQPatch",
                                             "areas", "links"))

##Read with alias
readAntares(select = "adqPatch", opts = opts, mcYears = 1)


}

}
