% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateResult.R
\name{aggregateResult}
\alias{aggregateResult}
\title{Creation of Mc_all}
\usage{
aggregateResult(opts, newname, verbose = 1)
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}}

\item{newname}{\code{character} name of simulation.}

\item{verbose}{\code{numeric} show log in console. Defaut to 1
\itemize{
 \item{0}{ : No log}
 \item{1}{ : Short log}
 \item{2}{ : Long log}
}}
}
\description{
Creation of Mc_all
}
