% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimB.R
\name{askProbleme}
\alias{askProbleme}
\title{Write optimisation probleme}
\usage{
askProbleme(pointX, faceY, face)
}
\arguments{
\item{pointX}{\code{data.table}, extreme points for 3 country, BE, DE anf FR}

\item{faceY}{\code{data.table}, face for 3 country, BE, DE anf FR for all tuple in face}

\item{face}{\code{data.table}, face for 3 country, BE, DE anf FR}
}
\description{
Write optimisation probleme
}
