% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFBTS.R
\name{createFBTS}
\alias{createFBTS}
\title{Create time series of typical days for antares}
\usage{
createFBTS(opts, probabilityMatrix, multiplier, interSeasonBegin,
  interSeasonEnd, firstDay, seed = 4052017, silent = FALSE)
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. Defaut to \code{antaresRead::simOptions()}}

\item{probabilityMatrix}{\code{list} from \code{flowBasedClustering::getProbability}. Columns must be rename to corespond to antares names.
Format used is : area_variable (Ex, fr_load, de_solar ...)}

\item{multiplier}{\code{data.frame} contain 2 columns.
\itemize{
\item variable : Name of variable
\item coef : coefficient multiplier.
}}

\item{interSeasonBegin}{\code{character or date}, date or vector of dates, YYYY-MM-DD, begin of interseason}

\item{interSeasonEnd}{\code{character or date}, date or vector of dates, YYYY-MM-DD, end of interseason}

\item{firstDay}{\code{numeric} indication of type of first day of study. If the first day of study is a
wednesday, you must specify firstDay = 3.}

\item{seed}{\code{numeric} fix random seed}

\item{silent}{\code{boolean} progress bar.}
}
\description{
Create time series of typical days for antares
}
\examples{

\dontrun{
library(antaresRead)
library(flowBasedClustering)
library(data.table)

# load climate daily time serires
climate <- fread(system.file("dataset/climate_example.txt",package = "flowBasedClustering"))

# load clustering results (or build them with clusteringTypicalDays function())
clusterTD <- readRDS(system.file("dataset/cluster_example.RDS",package = "flowBasedClustering"))

levelsProba <- list(summerWd = list(FR_load = c(0.5), DE_wind = c(1/3, 2/3), DE_solar = .5),
                    summerWe = list(FR_load = c(0.5, 0.7), DE_wind = c(.5)),
                    interSeasonWd = list(FR_load = c(0.5, 0.6),
                                         DE_wind = c(1/3, 2/3),
                                         DE_solar = 0.3), 
                    
                    interSeasonWe = list(FR_load = c(0.5, 0.6),
                                         DE_wind = c(1/3, 2/3),
                                         DE_solar = 0.3))
matProb <- getProbability(climate, clusterTD,
 levelsProba = levelsProba, extrapolationNA = TRUE)


opts <- antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/antaresStudy")



setNamesProbabilityMatrix(matProb, c("FR_load", "DE_wind", "DE_solar"),
                                   c("fr_load", "de_wind", "de_solar"))
setnames(matProb[[1]],"FR_load", "fr_load" )
setnames(matProb[[2]],"FR_load", "fr_load" )

setnames(matProb[[1]],"DE_wind", "de_wind" )
setnames(matProb[[2]],"DE_wind", "de_wind" )

setnames(matProb[[1]],"DE_solar", "de_solar" )
setnames(matProb[[2]],"DE_solar", "de_solar" )

multiplier <- data.frame(variable = c("fr_load", "de_wind", "de_solar"),
                         coef = c(1, 352250, 246403))
firstDay <- identifyFirstDay(opts)
interSeasonBegin <- as.Date(c("2017-09-03", "2018-02-02"))
interSeasonEnd <- as.Date(c("2017-10-04", "2018-05-02"))


ts <- createFBTS(opts = opts, probabilityMatrix = matProb, multiplier = multiplier,
                 interSeasonBegin = interSeasonBegin,
                 interSeasonEnd = interSeasonEnd, firstDay = firstDay)
}
                 
}
