% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{generateRaportFb}
\alias{generateRaportFb}
\title{Generate html report}
\usage{
generateRaportFb(allFB, hour, dayType)
}
\arguments{
\item{allFB}{\code{list}, object obtain which computeFB function}

\item{hour}{\code{numeric}, hour}

\item{dayType}{\code{numeric}, dayType}
}
\description{
Generate html report
}
