% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{generateReportFb}
\alias{generateReportFb}
\title{Generate html report}
\usage{
generateReportFb(dayType, output_file = NULL,
  fb_opts = antaresFlowbased::fbOptions(), allFB = NULL)
}
\arguments{
\item{dayType}{\code{numeric}, dayType}

\item{output_file}{\code{character}, output directory}

\item{fb_opts}{\code{list} of flowbased parameters returned by the function \link{setFlowbasedPath}. Defaut to \code{antaresFlowbased::fbOptions()}}

\item{allFB}{\code{data.table}, load FB directaly in package}
}
\description{
Generate html report
}
\examples{

\dontrun{
allFB <- computeFB(dayType = 7)
generateReportFb(dayType = 7, fb_opts = antaresFlowbased::fbOptions())
}
}
