% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifPtdf.R
\name{giveBClassif}
\alias{giveBClassif}
\title{Give B from PTDF}
\usage{
giveBClassif(PTDF, nbClust = 36)
}
\arguments{
\item{PTDF}{\code{data.frame}, PTDF}

\item{nbClust}{\code{numeric}, number of cluster}
}
\description{
Give B from PTDF
}
