% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimB.R
\name{giveTuples}
\alias{giveTuples}
\title{Gives tuples of B who check all constraints}
\usage{
giveTuples(face, pointX)
}
\arguments{
\item{face}{\code{data.table}, face for 3 country, BE, DE anf FR}

\item{pointX}{\code{data.table}, extreme points for 3 country, BE, DE anf FR}
}
\description{
Gives tuples of B who check all constraints
}
