% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graphFlowBased2D}
\alias{graphFlowBased2D}
\title{Plot 2D for flowbased areas}
\usage{
graphFlowBased2D(flowbased, ctry1, ctry2)
}
\arguments{
\item{flowbased}{\code{list}, flowbased outout obtain which computeFB function}

\item{ctry1}{\code{character}, country in X}

\item{ctry2}{\code{character}, country in Y}
}
\description{
Plot 2D for flowbased areas
}
