% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstDay.R
\name{identifyFirstDay}
\alias{identifyFirstDay}
\title{Find fist day of a study}
\usage{
identifyFirstDay(opts)
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. Defaut to \code{antaresRead::simOptions()}}
}
\description{
This function give type of first day, 1 is monday, 2 is Thuesday, ..., 7 is Sunday
}
\examples{

\dontrun{

opts <- setSimulationPath("MyPath")
identifyFirstDay(opts)

}

}
