% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_general_setting.R
\name{modifyGeneralSettingPlayList}
\alias{modifyGeneralSettingPlayList}
\title{Modify play list}
\usage{
modifyGeneralSettingPlayList(generaldataIniPatch, playList)
}
\arguments{
\item{generaldataIniPatch}{\code{character}, path of generaldataIni}

\item{playList}{\code{numeric}, playList}
}
