% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateResult.R
\name{moveFilesAfterStudy}
\alias{moveFilesAfterStudy}
\title{Move file after simulations}
\usage{
moveFilesAfterStudy(opts, simName)
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}}

\item{simName}{\code{character} name of simulation return by \link{runSimulation}}
}
\description{
Move file after simulations
}
\examples{

\dontrun{
}

}
