% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{plotFB}
\alias{plotFB}
\title{Plot 2D for flowbased areas}
\usage{
plotFB(dayType, hour, country1, country2,
  fb_opts = antaresFlowbased::fbOptions())
}
\arguments{
\item{dayType}{\code{numeric}, dayType}

\item{hour}{\code{numeric}, hour}

\item{flowbased}{\code{list}, flowbased outout obtain which computeFB function}

\item{ctry1}{\code{character}, country in X}

\item{ctry2}{\code{character}, country in Y}
}
\description{
Plot 2D for flowbased areas
}
