% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{positionViz}
\alias{positionViz}
\title{Graph function}
\usage{
positionViz(opts, fb_opts, dayType, hour, mcYears, ctry1, ctry2)
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. Defaut to \code{antaresRead::simOptions()}}

\item{fb_opts}{\code{list} of flowbased parameters returned by the function \link{setFlowbasedPath}. Defaut to \code{antaresFlowbased::fbOptions()}}

\item{dayType}{: day type}

\item{hour}{: hour}

\item{mcYears}{: mcYears}

\item{ctry1}{: first country}

\item{ctry2}{: second country}
}
\description{
Graph function
}
\examples{
study <- "D:/Users/titorobe/Desktop/exemple_test_BP"
# set the typical days used in the study
fb_opts <- setFlowbasedPath(model = "model2017")
# select the output study
output <- antaresRead::setSimulationPath(study, -1)

## plot a domain and the matching output points 
positionViz(opts = output, 
         fb_opts = fb_opts, 
         dayType = 5, hour = 19, 
         mcYears = 1, 
         ctry1 = "BE", ctry2 = "FR")
\dontrun{

}

}
