% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_simulation_files.R, R/runSimulation.R
\name{prepareSimulatioFiles}
\alias{prepareSimulatioFiles}
\alias{runSimulation}
\title{Preparation of sigle simulation files, write playlist and FB}
\usage{
prepareSimulatioFiles(opts, ts, second_membre, scenario, simNmber)

runSimulation(opts, simName, mc_all = TRUE, mc_ind = TRUE, indic = "all")
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}}

\item{ts}{\code{data.frame} of dayType and simulation}

\item{second_membre}{\code{data.frame} os second members}

\item{scenario}{\code{data.frame} of scenario}

\item{simNmber}{\code{numeric} numer of simulation must be in scenario}

\item{simName}{\code{wharacter} name of simulation}

\item{mc_all}{\code{boolean} give compress results}

\item{mc_ind}{\code{boolean} give mc_ind}

\item{indic}{\code{character} not use in this version}

\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}}
}
\description{
Preparation of sigle simulation files, write playlist and FB

Run all simulations
}
\examples{

\dontrun{
opts <- antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/exemple_test",0)
 second_membre <- data.table::fread(paste0(opts$studyPath,"/user/flowbased/second_member.txt"))
 ts <- data.table::fread(paste0(opts$studyPath,"/user/flowbased/ts.txt"))
 scenario <- data.table::fread(paste0(opts$studyPath,"/user/flowbased/scenario.txt"))
 prepareSimulatioFiles(opts = opts, ts = ts, second_membre = second_membre, scenario = scenario,
                       simNmber = 54 )
}


\dontrun{
opts <- antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/exemple_test",0)
mysim <- runSimulation(opts, "MystudyTest2")
}

}
