% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSimulationFiles.R
\name{prepareSimulationFiles}
\alias{prepareSimulationFiles}
\title{Prepare files needed for a single simulation run. Write playlist and FB}
\usage{
prepareSimulationFiles(ts, secondMember, scenarios, simNumber,
  opts = antaresRead::simOptions(), verbose = 1)
}
\arguments{
\item{ts}{\code{data.frame} of dayType and simulation}

\item{secondMember}{\code{data.frame} of second members}

\item{scenarios}{\code{data.frame} of scenarios}

\item{simNumber}{\code{numeric} number of simulation must be in scenarios}

\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. Defaut to \code{antaresRead::simOptions()}}

\item{verbose}{\code{numeric} show log in console. Defaut to 1
\itemize{
 \item{0}{ : No log}
 \item{1}{ : Short log}
 \item{2}{ : Long log}
}}
}
\description{
Prepare files needed for a single simulation run. Write playlist and FB
}
\examples{

\dontrun{

 path <- "D:/exemple_test"
 antaresRead::setSimulationPath(path, 0)

 secondMember <- data.table::fread(paste0(opts$studyPath,"/user/flowbased/second_member.txt"))
 ts <- data.table::fread(paste0(opts$studyPath,"/user/flowbased/ts.txt"))
 scenarios <- data.table::fread(paste0(opts$studyPath,"/user/flowbased/scenario.txt"))

 prepareSimulationFiles(opts = opts, ts = ts, secondMember = secondMember,
   scenarios = scenarios, simNumber = 54)
}

}
