% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimB.R
\name{resolvB}
\alias{resolvB}
\title{Resolve optimisation probleme}
\usage{
resolvB(pointX, faceY, probleme, alpha)
}
\arguments{
\item{pointX}{\code{data.table}, extreme points for 3 country, BE, DE anf FR}

\item{faceY}{\code{data.table}, face for 3 country, BE, DE anf FR for all tuple in face}

\item{probleme}{\code{optimization_model}, make which askProbleme}

\item{alpha}{\code{numeric}, between 0 and 1, error ponderation if 0 error of type 1 is ignored
if 1 error of type 0 is ignored}
}
\description{
Resolve optimisation probleme
}
