% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{runAppError}
\alias{runAppError}
\title{Run a shiny application to visualize the real and modelled flow-based domains}
\usage{
runAppError(fb_opts = antaresFlowbased::fbOptions())
}
\arguments{
\item{fb_opts}{\code{list} of flow-based parameters returned by the function \link{setFlowbasedPath}: directory of the flow-
based model to study.
By default, the value is indicated by \code{antaresFlowbased::fbOptions()}}
}
\description{
Run a shiny application displaying the results of the conversion of real domains into Antares models. It will display for
each typical day and each hour the volumetric errors of conversion (inf_error: forgotten points in the 
model, sup_error: modelled points missing in the real domain) and dynamic plots of the real and modelled domains. Html reports
for each day can also be exported.
}
\examples{

\dontrun{
 fb_opts = antaresFlowbased::fbOptions()
 runAppError(fb_opts)
}
}
