% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{runAppPosition}
\alias{runAppPosition}
\title{Run shiny visualisation of position}
\usage{
runAppPosition(dta, opts = antaresRead::simOptions())
}
\arguments{
\item{dta}{\code{antaresDataList} load with \code{antaresRead}}

\item{opts}{\code{simOptions} load with \code{setSimulationPath}}
}
\description{
Run shiny visualisation of position
}
\examples{

\dontrun{
opts <- antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/antaresStudy", -1)
dta <- antaresRead::readAntares(areas = c("fr", "be", "de", "nl"),
}

}
