% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimulation.R
\name{runSimulation}
\alias{runSimulation}
\title{Run all simulations}
\usage{
runSimulation(simulationName, mcAll = TRUE, mcInd = TRUE,
  indicators = c("mean", "min", "max", "sd"), .test = TRUE,
  MCyearsExcludes = NULL, silent = TRUE, opts = simOptions())
}
\arguments{
\item{simulationName}{\code{character} name of simulation}

\item{mcAll}{\code{boolean} give mc_all compress results}

\item{mcInd}{\code{boolean} keep mc_ind.}

\item{indicators}{\code{character} not use in this version}

\item{.test}{\code{boolean} if TRUE, just run 3 scenarios.}

\item{MCyearsExcludes}{\code{numeric} to exclude MCyearsExcludes, default NULL,
all MCyearsExcludes are included.}

\item{silent}{\code{boolean} show log in console.}

\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}}
}
\description{
Run all simulations
}
\examples{

\dontrun{
antaresRead::setSimulationPath("D:/exemple_test",0)

weight <- system.file("/test/data/coefficients_Antares.csv", package = "antaresFlowbased")
secondMember <- system.file("/test/data/fichier_b_final.csv", package = "antaresFlowbased")
dayType <- system.file("/test/data/id_FB.txt", package = "antaresFlowbased")

initFlowBased(weight = weight, secondMember = secondMember, dayType = dayType)

setSolverAntares(path = "C:\\\\Program Files\\\\RTE\\\\Antares\\\\5.0.9\\\\bin\\\\antares-5.0-solver.exe")

mysim <- runSimulation("R_from", silent = FALSE)
}

}
