% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimulation.R
\name{runSimulationFB}
\alias{runSimulationFB}
\title{Run antares Simulation with several flow-based time series}
\usage{
runSimulationFB(simulationName = "FlowBased", mcAll = TRUE, mcInd = TRUE,
  mcYears = "all", opts = antaresRead::simOptions(), verbose = 1,
  antaresParallel = TRUE)
}
\arguments{
\item{simulationName}{\code{character} name of simulation. Defaut to 'FlowBased'.}

\item{mcAll}{\code{boolean} give mc_all compress results. Defaut to TRUE.}

\item{mcInd}{\code{boolean} keep mc_ind. Defaut to TRUE.}

\item{mcYears}{\code{numeric} include mcYears. Default all (all mcYears are included)}

\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. Defaut to \code{antaresRead::simOptions()}}

\item{verbose}{\code{numeric} show log in console. Defaut to 1
\itemize{
 \item 0 : No log
 \item 1 : Short log
 \item 2 : Long log
}}

\item{antaresParallel}{\code{boolean} use antares parallel compute}
}
\description{
This function runs the simulation of an ANTARES study and changes the second members
of the binding constraints for each MC year so as to take into account several flow-based
domains time series in the model.

\code{runSimulationFB} function works on an ANTARES study whose input have been detailled with
four new files, located in user/flowbased/ directory :

\itemize{
 \item weight.txt : names and weights of the binding constraints which define the FB domains
 \item second_member.txt : second members of the binding constraints for each typical day and hour
 \item ts.txt : several time series of typical days to describe several aleas on flow-based domains
 \item scenario.txt : equivalent to scenario builder, define which time series should be use for each simulated mc-year
}

Those files can be automatically build with the function \link{initFlowBased}.
}
\examples{

\dontrun{
# set study
antaresRead::setSimulationPath("D:/exemple_test", 0)

# set solver path
setSolverAntares(path = "C:\\\\Program Files\\\\RTE\\\\Antares\\\\5.0.9\\\\bin\\\\antares-5.0-solver.exe")

# init directory for study
setFlowbasedPath(model = "model2017")
initFlowBased()

# run flowbased simulation
runSimulationFB(simulationName = "R_from", verbose = 1)
}

}
