% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antaresPath.R, R/ini_function.R
\name{setSolverAntares}
\alias{setSolverAntares}
\alias{getSolverAntares}
\alias{writeIni}
\title{Set the path of the antares solver executable}
\usage{

  setSolverAntares(path = "C:\\\\Program Files\\\\RTE\\\\Antares\\\\5.0.9\\\\bin\\\\antares-5.0-solver.exe")

getSolverAntares()

writeIni(listData, pathIni)
}
\arguments{
\item{path}{\code{Character}, Path to the executable antares solver}

\item{pathIni}{\code{Character}, Path to ini file}

\item{listData\code{list}, }{list otain by antaresRead:::readIniFile}
}
\description{
This function set and init an options "antares_solver" with path to the solver executable

Write ini file from list obtain by antaresRead:::readIniFile and modify by user
}
\examples{

\dontrun{

setSolverAntares(path = "C:\\\\Program Files\\\\RTE\\\\Antares\\\\5.0.9\\\\bin\\\\antares-5.0-solver.exe")

getSolverAntares()

}


\dontrun{
pathIni <- "D:/Users/titorobe/Desktop/exemple_test/settings/generaldata.ini"
generalSetting <- antaresRead:::readIniFile(pathIni)
generalSetting$output$synthesis <- FALSE
writeIni(generalSetting, pathIni)

}
}
