% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{updateBindingConstraintsIni}
\alias{updateBindingConstraintsIni}
\alias{changeBindingConstraints}
\title{Change bindingconstraints.ini file}
\usage{
updateBindingConstraintsIni(pathWeight, opts = antaresRead::simOptions())

changeBindingConstraints(pathWeight, opts = antaresRead::simOptions())
}
\arguments{
\item{pathWeight}{\code{Character}, Path to weight .txt file}

\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}}
}
\description{
Change bindingconstraints.ini file
}
\examples{

\dontrun{

updateBindingConstraintsIni(pathWeight <- "D:\\\\Users\\\\benothie\\\\Documents\\\\weigth.txt")

}

}
\seealso{
\code{writeBindingConstraintsIni}
}
