% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_general_setting.R
\name{updateGeneralSettingIni}
\alias{updateGeneralSettingIni}
\title{Change bindingconstraints.ini file}
\usage{
updateGeneralSettingIni(opts, playList = NULL)
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}}
}
\examples{

\dontrun{

opts <- antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/exemple_test",0)
updateGeneralSettingIni(opts)

}

}
\seealso{
\code{writeGeneralSettingIni}
}
