% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGeneralSetting.R
\name{updateGeneralSettingIni}
\alias{updateGeneralSettingIni}
\title{Change bindingconstraints.ini file}
\usage{
updateGeneralSettingIni(opts = antaresRead::simOptions(), playList = NULL)
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. Defaut to \code{antaresRead::simOptions()}}

\item{playList}{\code{numeric}, playList}
}
\examples{

\dontrun{

opts <- antaresRead::setSimulationPath("D:/exemple_test", 0)
updateGeneralSettingIni(opts)

}

}
\seealso{
\code{writeGeneralSettingIni}
}
