% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_constraints.R
\name{writeBindingConstraintsIni}
\alias{writeBindingConstraintsIni}
\title{Write bindingconstraints.ini file}
\usage{
writeBindingConstraintsIni(listData, opts = antaresRead::simOptions())
}
\arguments{
\item{listData}{\code{list}, bindingconstraints.ini as list R}

\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}}
}
\description{
Write bindingconstraints.ini file
}
\examples{

}
