% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_general_setting.R
\name{writeGeneralSettingIni}
\alias{writeGeneralSettingIni}
\title{Write generaldata.ini file}
\usage{
writeGeneralSettingIni(generaldataIniPatch, generalSetting)
}
\arguments{
\item{generaldataIniPatch}{\code{character} patch of generaldata.ini file}

\item{listData}{\code{list}, generaldata.ini as list R}
}
\description{
Write generaldata.ini file
}
