% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adqPatchH5.R
\name{adqH5}
\alias{adqH5}
\title{Run the adequacy patch and save result in h5}
\usage{
adqH5(opts, fb_opts, strategic_reserve_be = NULL,
  strategic_reserve_de = NULL)
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. Defaut to \code{antaresRead::simOptions()} must refer to a h5 file.}

\item{fb_opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath} or fb model localisation obtain with \link{setFlowbasedPath}. Defaut to \code{antaresRead::simOptions()}}

\item{strategic_reserve_be}{\code{character} area use to compute new margin for BE}

\item{strategic_reserve_de}{\code{character} area use to compute new margin for DE}
}
\description{
Compute adqPatch on h5 file
}
\examples{

\dontrun{


##Opts defind on txt files
opts <- antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/antaresStudy",2)

##Write h5 stud
antaresRead::writeAntaresH5()

##Do a setSimulationPath on h5 file
opts <- antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/Antares/antaresFlowbased",1)
fb_opts = antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/antaresStudy", 2)

##Compute adq on H5 file
adqH5(opts, fb_opts)


setAlias("adqPatch", "Alias for adqPatch", c("LOLD_ADQPatch",
                                             "UNSP. ENRG_ADQPatch", "DTG MRG_ADQPatch",
                                             "UNSP. ENRG_ADQPatch", "BALANCE_ADQPatch",
                                             "FLOW LIN._ADQPatch",
                                             "areas", "links"))

readAntares(select = "adqPatch", opts = opts, mcYears = 1)


}

}
