% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adqPatch.R
\name{adqPatch}
\alias{adqPatch}
\title{Run the adequacy patch}
\usage{
adqPatch(mcYears = "all", pre_filter = FALSE, strategic_reserve_be = NULL,
  strategic_reserve_de = NULL, opts = antaresRead::simOptions(),
  select = NULL)
}
\arguments{
\item{mcYears}{\code{numeric} include mcYears. Default all (all mcYears are load)}

\item{pre_filter}{\code{boolean} filter mcYears before adqPatch apply, load annual data and if LOLD>0 on annual data dont load this mcYears}

\item{strategic_reserve_be}{\code{character} area use to compute new margin for BE}

\item{strategic_reserve_de}{\code{character} area use to compute new margin for DE}

\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. Defaut to \code{antaresRead::simOptions()}}

\item{select}{\code{character}, columns to select (columns need for adqPatch are automaticaly add)}
}
\description{
Compute adqPatch for antares study
}
\examples{

\dontrun{
antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/exemple_test_BP", 2)

#No strategic reserve
res <- adqPatch()

#Strategic reserve
res <- adqPatch(strategic_reserve_de = "lu_de", strategic_reserve_be = "lu_be")

#Add a new column
res <- adqPatch(strategic_reserve_de = "lu_de", strategic_reserve_be = "lu_be", select = "COAL")

}

}
