% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateResult.R
\name{aggregateResult}
\alias{aggregateResult}
\title{Creation of Mc_all}
\usage{
aggregateResult(opts, newname, silent = TRUE)
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}}

\item{newname}{\code{character} name of simulation.}

\item{silent}{\code{boolean} show log in console.}
}
\description{
Creation of Mc_all
}
\examples{

\dontrun{
}

}
