% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFB.R
\name{computeFB}
\alias{computeFB}
\title{Compute flowbased approximation}
\usage{
computeFB(PTDF = system.file("/input/ptdf/PTDF.csv", package =
  "antaresFlowbased"), outputName = paste0(getwd(), "/antaresInput"),
  reports = TRUE, dayType = "All", hour = "All", nbFaces = 36,
  verbose = 0)
}
\arguments{
\item{PTDF}{\code{character}, path for PTDF file format : 
\itemize{
 \item Id_day : column of id day, numeric. In default exemple between 1 and 12. 
 \item Period : column of period, numeric. In default exemple between 1 and 24.
 \item BE : column of BE, numeric. 
 \item DE : column of DE, numeric. 
 \item FR : column of FR, numeric. 
 \item FR : column of FR, numeric. 
 \item RAM : column of RAM, numeric. 
}}

\item{outputName}{\code{character}, name of output directory}

\item{reports}{\code{boolean}, product html reports by typical day or not}

\item{dayType}{\code{character / numeric} default All, can specify dayType to compute}

\item{hour}{\code{character / numeric} default All, can specify hour to compute}

\item{nbFaces}{\code{numeric} number of faces to keep, default 36.}

\item{verbose}{\code{numeric} show log in console. Defaut to 0.
\itemize{
 \item 0 : No log
 \item 1 : Short log
}}
}
\description{
Compute flowbased approximation
}
