% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFB.R
\name{computeFB}
\alias{computeFB}
\title{Compute flowbased approximation}
\usage{
computeFB(PTDF = system.file("/input/ptdf/PTDF.csv", package =
  "antaresFlowbased"), outputName = paste0(getwd(), "/antaresInput"),
  reports = TRUE, dayType = "All", hour = "All", nbFaces = 36)
}
\arguments{
\item{PTDF}{\code{character}, path for PTDF file}

\item{outputName}{\code{character}, name of output directory}

\item{reports}{\code{boolean}, product html reports by typical day or not}

\item{dayType}{\code{character / numeric} default All, can specify dayType to compute}

\item{hour}{\code{character / numeric} default All, can specify hour to compute}

\item{nbFaces}{\code{numeric} number of faces to keep, default 36.}
}
\description{
Compute flowbased approximation
}
