% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFB.R
\name{computeFB}
\alias{computeFB}
\title{Conversion of flow-based domains into an ANtares model with fixed PTDF}
\usage{
computeFB(PTDF = system.file("/input/ptdf/PTDF.csv", package =
  "antaresFlowbased"), outputName = paste0(getwd(), "/antaresInput"),
  reports = TRUE, dayType = "All", hour = "All", nbFaces = 36,
  verbose = 0)
}
\arguments{
\item{PTDF}{\code{character}, path leading to the flow-based domains (PTDF description) list. 
By default, this leads to an example file (\code{"PTDF.csv"}, saved in the package).
 This must be a csv file containing the following columns (and column names): 
\itemize{
 \item Id_day : numeric, name of each day. Default in example id_day between 1 and 12. 
 \item Period : numeric, hour of the day. Default in example period between 1 and 24 (1 is then between 00:00 and 01:00).
 \item BE : numeric, PTDF coefficient of Belgium.
 \item DE : numeric, PTDF coefficient of Germany. 
 \item FR : numeric, PTDF coefficient of France. 
 \item NL : numeric, PTDF coefficient of the Netherlands. 
 \item RAM : numeric, remaining available capacity in the critical branch (MW). 
}}

\item{outputName}{\code{character}, path/name of the output directory. By default, the value is a created directory named 
"antaresInput" in the current directory}

\item{reports}{\code{boolean}, if TRUE, the function will write html reports (one per typical day). By default, the value is 
TRUE.}

\item{dayType}{\code{numeric}, by default, the value is All. (optional) Vector of id_days to compute.}

\item{hour}{\code{numeric}, by default, the value is All. (optional) vector of hours/periods to compute.}

\item{nbFaces}{\code{numeric}, standard shape parameters: number of sides to select. By default, the value is 36.}

\item{verbose}{\code{numeric}, shows the logs in console. By default, the value is 0.
\itemize{
 \item 0 : No log
 \item 1 : Short log
}}
}
\description{
This function follows two steps: based on the list of flow-based domains given in input, it will calculate a standard shape
 (selection of sides, k-medoid method) and will then project the real domains on this standard shape. The projection is 
computed using an optimisation algorithm aiming at minimising the volumetric error between the real domain and its projection.
 The function will then write in an output directory the standard shape (weights.txt), the projection result for each domain 
 (second_member.txt) and an RDS object storing information on the projection and the errors. Reports can be written represening the real and
modelled domains and the volumetric error between them.
}
\examples{
\dontrun{
#Compute models for all days and hours of a PTDF file, with no reports automatically generated at the same time
computeFB(PTDF = system.file("/input/ptdf/PTDF.csv", package = "antaresFlowbased"), reports = FALSE)

}
}
