% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFB.R
\name{cumputeFB}
\alias{cumputeFB}
\title{Compute flowbased approximation}
\usage{
cumputeFB(PTDF = system.file("/optimWork/PTDF.csv", package =
  "antaresFlowbased"), face = system.file("/optimWork/B.csv", package =
  "antaresFlowbased"), dayType = "All", hour = "All")
}
\arguments{
\item{PTDF}{\code{character}, path for PTDF file}

\item{face}{\code{character}, path for face file}

\item{dayType}{\code{character / numeric} default All, can specify dayType to compute}

\item{hour}{\code{character / numeric} default All, can specify hour to compute}

\item{constrain}{\code{character}, path for constrain file}
}
\description{
Compute flowbased approximation
}
