% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFlowbasedPath.R
\name{flowbased-path}
\alias{flowbased-path}
\alias{setFlowbasedPath}
\alias{fbOptions}
\alias{getAvailabledBP}
\title{Set of function to control the flowbased input repository}
\usage{
setFlowbasedPath(path, bp)

fbOptions()

getAvailabledBP()
}
\arguments{
\item{path}{(optional)
If "bp" is missing. Path to the input repertory. . Must have :
\itemize{
 \item{coefficients_Antares.csv}{}
 \item{fichier_b_final.csv}{}
 \item{id_FB.csv}{}
 \item{domainesFB.RDS}{}
}}

\item{bp}{(optional) If "path" is missing. The name of input availabled inside the package. Linked to \code{\link{getAvailabledBP}}}
}
\value{
A vector of availabled bp for \code{getAvailabledBP}. For \code{setFlowbasedPath} and \code{fbOptions}, a list containing :
  \item{path}{path of the current input repository}
}
\description{
Set of function to control the flowbased input repository
}
\details{
The repertory choosen with \code{setFlowbasedPath} becomes the default
parameters for all functions of the package. The package have some availabled bp \code{getAvailabledBP}.
Current path can be retrieve using \code{fbOptions}
}
\examples{

\dontrun{
# Defaut path set loading the package

fbOptions()

# Specify a availabled bp
getAvailabledBP()
setFlowbasedPath(bp = "BP2017")

# Select a repository
setFlowbasedPath(bp = "C:/PATH/TO/INPUT")
}

}
