% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFlowbasedPath.R
\name{flowbased-path}
\alias{flowbased-path}
\alias{setFlowbasedPath}
\alias{fbOptions}
\alias{getAvailableModel}
\title{Set of function to control the flowbased input repository}
\usage{
setFlowbasedPath(path, model)

fbOptions()

getAvailableModel()
}
\arguments{
\item{path}{(optional)
If "model" is missing. Path to the input repertory. . Must have :
\itemize{
 \item{weight.txt}{}
 \item{second_member.txt}{}
 \item{ts.txt}{}
 \item{domainesFB.RDS}{}
}}

\item{model}{(optional) If "path" is missing. The name of input available inside the package. Linked to \code{\link{getAvailableModel}}}
}
\value{
A vector of available model for \code{getAvailableModel}. For \code{setFlowbasedPath} and \code{fbOptions}, a list containing :
  \item{path}{path of the current input repository}
}
\description{
Set of function to control the flowbased input repository
}
\details{
The repertory choosen with \code{setFlowbasedPath} becomes the default
parameters for all functions of the package. The package have some availabled model \code{getAvailableModel}.
Current path can be retrieve using \code{fbOptions}
}
\examples{

\dontrun{
# Defaut path set loading the package
fbOptions()

# Specify a available model
getAvailableModel()
setFlowbasedPath(model = "model2017")

# Select a repository
setFlowbasedPath(model = "C:/PATH/TO/INPUT")
}

}
