% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{generateRaportFb}
\alias{generateRaportFb}
\title{Generate html report}
\usage{
generateRaportFb(allFB, dayType, output_file = NULL)
}
\arguments{
\item{allFB}{\code{list}, object obtain which computeFB function}

\item{dayType}{\code{numeric}, dayType}
}
\description{
Generate html report
}
\examples{

\dontrun{
allFB <- computeFB(dayType = 7)
generateRaportFb(allFB, dayType = 7)
}
}
