% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{generateRaportFb}
\alias{generateRaportFb}
\title{Generate html report}
\usage{
generateRaportFb(dayType, output_file = NULL,
  fb_opts = antaresFlowbased::fbOptions())
}
\arguments{
\item{dayType}{\code{numeric}, dayType}

\item{output_file}{\code{character}, output directory}

\item{fb_opts}{\code{list} of flowbased parameters returned by the function \link{setFlowbasedPath}. Defaut to \code{antaresFlowbased::fbOptions()}}
}
\description{
Generate html report
}
\examples{

\dontrun{
allFB <- computeFB(dayType = 7)
generateRaportFb(allFB, dayType = 7)
}
}
