% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstDay.R
\name{identifyFirstDay}
\alias{identifyFirstDay}
\title{Find first day of a study}
\usage{
identifyFirstDay(opts, firstArea = "fr", secondArea = c("fr", "de", "be",
  "lu", "nl"))
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. Defaut to \code{antaresRead::simOptions()}}

\item{firstArea}{\code{character} first area(s) use to compute first day}

\item{secondArea}{\code{character} second area(s) use to compute first day, it's a security when you want to
be sure of first day compute. It can be NULL but your compute will be less sure.}
}
\description{
This function give type of first day, 1 is monday, 2 is Thuesday, ..., 7 is Sunday
}
\examples{

\dontrun{

opts <- setSimulationPath("MyPath")
identifyFirstDay(opts)

}

}
