% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initFlowBased.R
\name{initFlowBased}
\alias{initFlowBased}
\title{Generate environnement for a flow-based study}
\usage{
initFlowBased(opts, weigth, secondMember, dayType = NULL,
  determinants = NULL)
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}}

\item{weigth}{\code{character}, path of weigth file}

\item{secondMember}{\code{character}, path of secondMember file}

\item{dayType}{\code{character}, path of dayType file}

\item{determinants}{\code{character}, vector of determinants, not use in this version}
}
\description{
Generate environnement for a flow-based study
}
\examples{
\dontrun{
opts <- antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/exemple_test",0)
weigth <- "D:/Users/titorobe/Desktop/run_antares_FB/Write chroniques/Coefficients_Antares.csv"
secondMember <- "D:/Users/titorobe/Desktop/run_antares_FB/Write chroniques/fichier b final.csv"
dayType <- "D:/Users/titorobe/Desktop/run_antares_FB/Write chroniques/id_FB.txt"
initFlowBased(opts = opts,
              weigth = weigth,
              secondMember = secondMember,
              dayType = dayType)
}
}
