% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initFlowBased.R
\name{initFlowBased}
\alias{initFlowBased}
\title{Generate environment for a flow-based study}
\usage{
initFlowBased(weight = system.file("/input/coefficients_Antares.csv", package
  = "antaresFlowbased"), secondMember = secondMember <-
  system.file("/input/fichier_b_final.csv", package = "antaresFlowbased"),
  dayType = system.file("/input/id_FB.txt", package = "antaresFlowbased"),
  opts = antaresRead::simOptions())
}
\arguments{
\item{weight}{\code{character}, path of weight file. Defaut use data in package.}

\item{secondMember}{\code{character}, path of secondMember file. Defaut use data in package.}

\item{dayType}{\code{character}, path of dayType file. Defaut use data in package.}

\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. Defaut to \code{antaresRead::simOptions()}}
}
\description{
Generate environment for a flow-based study
}
\examples{

\dontrun{
# target antares study
path <- "D:/exemple_test"
antaresRead::setSimulationPath(path, 0)

initFlowBased()
}

}
