% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initFlowBased.R
\name{initFlowBased}
\alias{initFlowBased}
\title{Generate environment for a flow-based study}
\usage{
initFlowBased(fb_opts = antaresFlowbased::fbOptions(),
  opts = antaresRead::simOptions(), scenarios = rep(1:200, times = 5))
}
\arguments{
\item{fb_opts}{\code{list} of flowbased parameters returned by the function \link{setFlowbasedPath}. Defaut to \code{antaresFlowbased::fbOptions()}}

\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. Defaut to \code{antaresRead::simOptions()}}

\item{scenarios}{\code{numeric} scenarios use for write scenario.txt.}
}
\description{
Generate environment for a flow-based study
}
\note{
folder deigned by fb_opts contain files :
\itemize{
  \item{domainesFB.RDS}{RDS file from \link{computeFB}}
  \item{second_member.txt}{txt file wich following columns
  \itemize{
    \item{Id_day : numeric from 1 to number of day id}
    \item{Id_hour : numeric from 1 to number of hour}
    \item{vect_b : numeric}
    \item{Name : character, name of constaints}
  }}
 }
}
\examples{

\dontrun{
# target antares study
path <- "D:/exemple_test"
antaresRead::setSimulationPath(path, 0)

setFlowbasedPath(model = "model2017")

initFlowBased()
}

}
