% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initFlowBased.R
\name{initFlowBased}
\alias{initFlowBased}
\title{Generate environnement for a flow-based study}
\usage{
initFlowBased(weight = system.file("/data/coefficients_Antares.csv", package =
  "antaresFlowbased"), secondMember = secondMember <-
  system.file("/data/fichier_b_final.csv", package = "antaresFlowbased"),
  dayType = system.file("/data/id_FB.txt", package = "antaresFlowbased"),
  opts = antaresRead::simOptions())
}
\arguments{
\item{weight}{\code{character}, path of weight file}

\item{secondMember}{\code{character}, path of secondMember file}

\item{dayType}{\code{character}, path of dayType file}

\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. Defaut to \code{antaresRead::simOptions()}}
}
\description{
Generate environnement for a flow-based study
}
\examples{
\dontrun{

path <- "D:/exemple_test"
antaresRead::setSimulationPath(path, 0)

initFlowBased()
}

}
