% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initFlowBased.R
\name{initFlowBased}
\alias{initFlowBased}
\title{Generate environment for a flow-based study}
\usage{
initFlowBased(fb_opts = antaresFlowbased::fbOptions()$path,
  opts = antaresRead::simOptions(), scenarios = rep(1:200, times = 5),
  controlAntares = TRUE)
}
\arguments{
\item{fb_opts}{\code{list} of flowbased parameters returned by the function \link{setFlowbasedPath}. Defaut to \code{antaresFlowbased::fbOptions()}}

\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. Defaut to \code{antaresRead::simOptions()}}

\item{scenarios}{\code{numeric} scenarios use for write scenario.txt.}

\item{controlAntares}{\code{boolean} for test.}
}
\description{
Generate environment for a flow-based study
}
\note{
folder deigned by fb_opts contain files :
\itemize{
  \item{domainesFB.RDS}{RDS file from \link{computeFB}}
  \item{second_member.txt}{txt file of second member wich following columns :
  \itemize{
    \item{Id_day : numeric from 1 to number of day id}
    \item{Id_hour : numeric from 1 to number of hour}
    \item{vect_b : numeric}
    \item{Name : character, name of constaints}
  }}
  \item{ts.txt}{txt file of time series matrix wich
  \itemize{
    \item{In row : dates, format : %YYYY-%MM-%DD
    \item{In column : vector of chronics}
    \item{in cell, numeric (typival day ID)}
  }}}
  \item{weigth.txt}{weigth file  wich following columns :
  \itemize{
    \item{Name : character, name of contraint}
    \item{BE.FR : numeric, between -1 and 1}
    \item{DE.FR : numeric, between -1 and 1}
    \item{DE.NL : numeric, between -1 and 1}
    \item{BE.NL : numeric, between -1 and 1}
    \item{BE.DE : numeric, between -1 and 1}
  }}}
}
\examples{

\dontrun{

 antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/antaresStudy")
 setSolverAntares("D:/Users/titorobe/Desktop/RTAANTARES/bin/antares-6.1-solver.exe")
 initFlowBased()
 }
 
 
}
