% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initFlowBased.R
\name{initFlowBased}
\alias{initFlowBased}
\title{Generate environnement for a flow-based study}
\usage{
initFlowBased(opts, weigth, secondMember, dayType = NULL,
  determinants = NULL)
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}}

\item{weigth}{\code{character}, path of weigth file}

\item{secondMember}{\code{character}, path of secondMember file}

\item{dayType}{\code{character}, path of dayType file}

\item{determinants}{\code{character}, vector of determinants, not use in this version}
}
\description{
Generate environnement for a flow-based study
}
\examples{
\dontrun{

path <- "D:/Users/titorobe/Desktop/exemple_test"
opts <- antaresRead::setSimulationPath(path, 0)

weigth <- system.file("/test/data/coefficients_Antares.csv", package = "antaresFlowbased")
secondMember <- system.file("/test/data/fichier_b_final.csv", package = "antaresFlowbased")
dayType <- system.file("/test/data/id_FB.txt", package = "antaresFlowbased")

initFlowBased(opts = opts,
              weigth = weigth,
              secondMember = secondMember,
              dayType = dayType)
}

}
