% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_general_setting.R
\name{modifyGeneralSetting}
\alias{modifyGeneralSetting}
\title{update general data}
\usage{
modifyGeneralSetting(generaldataIniPatch)
}
\arguments{
\item{generaldataIniPatch}{\code{character}, path of generaldataIni}
}
