% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGeneralSetting.R
\name{modifyGeneralSettingPlayList}
\alias{modifyGeneralSettingPlayList}
\title{Modify playList}
\usage{
modifyGeneralSettingPlayList(generaldataIniPatch, playList)
}
\arguments{
\item{generaldataIniPatch}{\code{character}, path of generaldataIni}

\item{playList}{\code{numeric}, playList}
}
\description{
Modify playList
}
