% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateResult.R
\name{moveFilesAfterStudy}
\alias{moveFilesAfterStudy}
\title{Move file after simulations}
\usage{
moveFilesAfterStudy(opts, simulationName, verbose = 1)
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}}

\item{simulationName}{\code{character} name of simulation return by \link{runSimulation}}

\item{verbose}{\code{numeric} show log in console. Defaut to 1
\itemize{
 \item{0}{ : No log}
 \item{1}{ : Short log}
 \item{2}{ : Long log}
}}
}
\description{
Move file after simulations
}
\examples{

}
