% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{plotFB}
\alias{plotFB}
\title{Plot 2D for flowbased areas can combine more than one plot}
\usage{
plotFB(dayType, hour, country1, country2,
  fb_opts = antaresFlowbased::fbOptions())
}
\arguments{
\item{dayType}{\code{numeric}, dayType}

\item{hour}{\code{numeric}, hour}

\item{country1}{\code{character}, country in X}

\item{country2}{\code{character}, country in Y}

\item{fb_opts}{\code{list} of flowbased parameters returned by the function \link{setFlowbasedPath}. Defaut to \code{antaresFlowbased::fbOptions()}}
}
\description{
Plot 2D for flowbased areas can combine more than one plot
}
\examples{

\dontrun{
 fb_opts = antaresFlowbased::fbOptions()
 plotFB(dayType = 1, hour = 1, country1 = "FR", country2 = "NL", fb_opts = fb_opts)
 plotFB(dayType = 1:2, hour = 1,country1 = "FR",country2 = "NL", fb_opts = fb_opts)
 plotFB(dayType = 1:2, hour = 1:2, country1 = "FR", country2 = "NL", fb_opts = fb_opts)
 plotFB(dayType = 1, hour = 1, country1 = c("FR", "DE"), country2 = c("NL", "FR"), fb_opts = fb_opts)
}


}
