% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{plotFB}
\alias{plotFB}
\title{Plot 2D for flowbased areas}
\usage{
plotFB(dayType, hour, country1, country2,
  fb_opts = antaresFlowbased::fbOptions())
}
\arguments{
\item{dayType}{\code{numeric}, dayType}

\item{hour}{\code{numeric}, hour}

\item{country1}{\code{character}, country in X}

\item{country2}{\code{character}, country in Y}

\item{fb_opts}{\code{list} of flowbased parameters returned by the function \link{setFlowbasedPath}. Defaut to \code{antaresFlowbased::fbOptions()}}
}
\description{
Plot 2D for flowbased areas
}
\examples{

\dontrun{
plotFB(1,1,"FR","NL")
plotFB(1:2,1,"FR","NL")
plotFB(1:2,1:2,"FR","NL")
plotFB(1,1,c("FR", "DE"),c("NL", "FR"))
}


}
