% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{positionViz}
\alias{positionViz}
\title{Graph function}
\usage{
positionViz(opts, data, dayType, hour, ctry1, ctry2,
  filteringEmptyDomains = FALSE, nbMaxPt = 10000)
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. Defaut to \code{antaresRead::simOptions()}}

\item{data}{\code{antaresDataList} import with \link{readAntares}}

\item{dayType}{: day type, can be numeric or 'all'}

\item{hour}{: hour, can be numeric or 'all'}

\item{ctry1}{: first country}

\item{ctry2}{: second country}

\item{filteringEmptyDomains}{\code{boolean} filtering empty domains}

\item{nbMaxPt}{\code{numeric} number of point maximum on graph. Default 10000.}
}
\description{
Graph function
}
\examples{
\dontrun{
study <- "D:/Users/titorobe/Desktop/antaresStudy"

opts <- antaresRead::setSimulationPath(study, -1)
dta <- antaresRead::readAntares(areas = c("fr", "be", "de", "nl"), 
                                links = c("be - de","be - fr","be - nl",
                                "de - fr","de - nl"), mcYears = 1:10,
                                select = c("LOLD", "UNSP. ENRG", 
                                "DTG MRG", "UNSP. ENRG", "BALANCE", "FLOW LIN."),
                                 opts = opts)

## plot a domain and the matching output points 
positionViz(opts = opts, 
         data = dta,
         dayType = 1, hour = 19:20, 
         ctry1 = "BE", ctry2 = "FR")
         
dta$areas <- dta$areas[timeId == 1]
## plot a sigle idTime with all domains 
positionViz(opts = opts, 
         data = dta,
         dayType = "all", hour = 0, 
         ctry1 = "BE", ctry2 = "FR")
         
##Filtering empty domains

positionViz(opts = opts, 
         data = dta,
         dayType = "all", hour = 0, 
         ctry1 = "BE", ctry2 = "FR", filteringEmptyDomains = TRUE)
}

}
