% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{positionViz}
\alias{positionViz}
\title{Graph function}
\usage{
positionViz(data, dayType, hour, country1, country2,
  fb_opts = antaresRead::simOptions(), filteringEmptyDomains = FALSE,
  nbMaxPt = 10000)
}
\arguments{
\item{data}{\code{antaresDataList} import with \link{readAntares}}

\item{dayType}{: day type, can be numeric or 'all'}

\item{hour}{: hour, can be numeric or 'all'}

\item{country1}{: first country}

\item{country2}{: second country}

\item{fb_opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath} or fb model localisation obtain with \link{setFlowbasedPath}. Defaut to \code{antaresRead::simOptions()}}

\item{filteringEmptyDomains}{\code{boolean} filtering empty domains}

\item{nbMaxPt}{\code{numeric} number of point maximum on graph. Default 10000.}
}
\description{
Graph function
}
\examples{
\dontrun{
study <- "D:/Users/titorobe/Desktop/antaresStudy"

fb_opts <- antaresRead::setSimulationPath(study, -1)
dta <- antaresRead::readAntares(areas = c("fr", "be", "de", "nl"), 
                                links = c("be - de","be - fr","be - nl",
                                "de - fr","de - nl"), mcYears = 1:10,
                                select = c("LOLD", "UNSP. ENRG", 
                                "DTG MRG", "UNSP. ENRG", "BALANCE", "FLOW LIN."),
                                 opts = opts)

## plot a domain and the matching output points 
positionViz(fb_opts = fb_opts, 
         data = dta,
         dayType = 1, hour = c(9, 19), 
         country1 = "BE", country2 = "FR")
         
dta$areas <- dta$areas[timeId == 1]
## plot a sigle idTime with all domains 
positionViz(fb_opts = fb_opts, 
         data = dta,
         dayType = "all", hour = 0, 
         country1 = "BE", country2 = "FR")
         
##Filtering empty domains

positionViz(fb_opts = fb_opts, 
         data = dta,
         dayType = "all", hour = 0, 
         country1 = "BE", country2 = "FR", filteringEmptyDomains = TRUE)
}

}
