% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{runAppError}
\alias{runAppError}
\title{Run shiny visualisation of error}
\usage{
runAppError(fb_opts = antaresFlowbased::fbOptions())
}
\arguments{
\item{fb_opts}{\code{list} of flowbased parameters returned by the function \link{setFlowbasedPath}. Defaut to \code{antaresFlowbased::fbOptions()}}
}
\description{
Run shiny visualisation of error
}
