% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{runAppPosition}
\alias{runAppPosition}
\title{Run shiny visualisation of position}
\usage{
runAppPosition(dta, fb_opts = antaresRead::simOptions())
}
\arguments{
\item{dta}{\code{antaresDataList} load with \code{antaresRead}}

\item{fb_opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath} or fb model localisation obtain with \link{setFlowbasedPath}. Defaut to \code{antaresRead::simOptions()}}
}
\description{
Run shiny visualisation of position
}
\examples{

\dontrun{
opts <- antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/antaresStudy", -1)
dta <- antaresRead::readAntares(areas = c("fr", "be", "de", "nl"),
                                links = c("be - de","be - fr","be - nl","de - fr","de - nl"), mcYears = 1:10,
                                select = c("LOLD", "UNSP. ENRG", "DTG MRG", "UNSP. ENRG", "BALANCE", "FLOW LIN."), opts = opts)
runAppPosition(dta)
}

}
