% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimulation.R
\name{runSimulation}
\alias{runSimulation}
\title{Run all simulations}
\usage{
runSimulation(opts, simulationName, mcAll = TRUE, mcInd = TRUE,
  indicators = c("mean", "min", "max", "sd"), .test = TRUE)
}
\arguments{
\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}}

\item{simulationName}{\code{character} name of simulation}

\item{mcAll}{\code{boolean} give mc_all compress results}

\item{mcInd}{\code{boolean} keep mc_ind.}

\item{indicators}{\code{character} not use in this version}

\item{.test}{\code{boolean} if TRUE, just run 3 scenarios.}
}
\description{
Run all simulations
}
\examples{

\dontrun{
opts <- antaresRead::setSimulationPath("D:/Users/titorobe/Desktop/exemple_test",0)

weigth <- system.file("/test/data/coefficients_Antares.csv", package = "antaresFlowbased")
secondMember <- system.file("/test/data/fichier_b_final.csv", package = "antaresFlowbased")
dayType <- system.file("/test/data/id_FB.txt", package = "antaresFlowbased")

initFlowBased(opts = opts, weigth = weigth, secondMember = secondMember, dayType = dayType)

setSolverAntares(path = "C:\\\\Program Files\\\\RTE\\\\Antares\\\\5.0.9\\\\bin\\\\antares-5.0-solver.exe")

mysim <- runSimulation(opts, "MystudyTest2")
}

}
