% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimulation.R
\name{runSimulation}
\alias{runSimulation}
\title{Run all simulations}
\usage{
runSimulation(simulationName = "FlowBased", mcAll = TRUE, mcInd = TRUE,
  mcYears = NULL, opts = antaresRead::simOptions(), silent = TRUE,
  .test = TRUE)
}
\arguments{
\item{simulationName}{\code{character} name of simulation}

\item{mcAll}{\code{boolean} give mc_all compress results}

\item{mcInd}{\code{boolean} keep mc_ind.}

\item{mcYears}{\code{numeric} include mcYears, default NULL,
all mcYears are included.}

\item{opts}{\code{list} of simulation parameters returned by the function \link{setSimulationPath}. Defaut to \code{antaresRead::simOptions()}}

\item{silent}{\code{boolean} show log in console.}

\item{.test}{\code{boolean} if TRUE, just run 3 scenarios.}
}
\description{
Run all simulations
}
\examples{

\dontrun{
antaresRead::setSimulationPath("D:/exemple_test",0)

initFlowBased()

setSolverAntares(path = "C:\\\\Program Files\\\\RTE\\\\Antares\\\\5.0.9\\\\bin\\\\antares-5.0-solver.exe")

mysim <- runSimulation("R_from", silent = FALSE)
}

}
