% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimBMatrix.R
\name{searchAlpha}
\alias{searchAlpha}
\title{Search alpha for a daytype and a hour}
\usage{
searchAlpha(face, pointX, faceY, probleme, PTDF, univ)
}
\arguments{
\item{face}{\code{data.table}, face for 3 country, BE, DE anf FR}

\item{pointX}{\code{data.table}, extreme points for 3 country, BE, DE anf FR}

\item{faceY}{\code{data.table}, face for 3 country, BE, DE anf FR for all tuple in face}

\item{probleme}{\code{optimization_model}, make which askProbleme}

\item{PTDF}{\code{data.frame} PTDF}

\item{univ}{\code{matrix} generate which .univ}
}
\description{
Search alpha for a daytype and a hour
}
