% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFBTS.R
\name{setNamesProbabilityMatrix}
\alias{setNamesProbabilityMatrix}
\title{Change name for probabilityMatrix}
\usage{
setNamesProbabilityMatrix(data, oldName, newName)
}
\arguments{
\item{data}{\code{list} from \code{flowBasedClustering::getProbability}. Columns must be rename to corespond to antares names.}

\item{oldName}{\code{character} vector of old names}

\item{newName}{\code{character} vector of new names}
}
\description{
Change name for probabilityMatrix
}
