% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antaresPath.R
\name{setSolverAntares}
\alias{setSolverAntares}
\alias{getSolverAntares}
\title{Set the path of the antares solver executable}
\usage{

  setSolverAntares(path = "C:/Program Files/RTE/Antares/5.0.9/bin/antares-5.0-solver.exe")

getSolverAntares()
}
\arguments{
\item{path}{\code{Character}, Path to the executable antares solver}
}
\description{
This function set and init an options "antares_solver" with path to the solver executable
}
\examples{

\dontrun{

setSolverAntares(path = "antares-solver.exe")

getSolverAntares()

}

}
