% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGeneralSetting.R
\name{writeGeneralSettingIni}
\alias{writeGeneralSettingIni}
\title{Write generaldata.ini file}
\usage{
writeGeneralSettingIni(generaldataIniPatch, generalSetting)
}
\arguments{
\item{generaldataIniPatch}{\code{list}}

\item{generalSetting}{\code{Character}, Path to ini file}
}
\description{
Write generaldata.ini file
}
