% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ini_function.R
\name{writeIni}
\alias{writeIni}
\title{Write ini file from list obtain by antaresRead:::readIniFile and modify by user}
\usage{
writeIni(listData, pathIni)
}
\arguments{
\item{listData}{\code{list}, modified list otain by antaresRead:::readIniFile}

\item{pathIni}{\code{Character}, Path to ini file}
}
\description{
Write ini file from list obtain by antaresRead:::readIniFile and modify by user
}
\examples{

\dontrun{
pathIni <- "D:/Users/titorobe/Desktop/exemple_test/settings/generaldata.ini"
generalSetting <- antaresRead:::readIniFile(pathIni)
generalSetting$output$synthesis <- FALSE
writeIni(generalSetting, pathIni)
}

}
